# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for dev.",
        "title": "dtle API Docs",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v2/validation/job": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "validate job config.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "validation"
                ],
                "operationId": "ValidateJobV2",
                "parameters": [
                    {
                        "description": "validate job config",
                        "name": "job_config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ValidateJobReqV2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ValidateJobRespV2"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.BinlogValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is a string version of any error that may have occured",
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "models.ConnectionValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is a string version of any error that may have occured",
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "models.DataSourceConfig": {
            "type": "object",
            "properties": {
                "table_schema": {
                    "type": "string"
                },
                "table_schema_regex": {
                    "type": "string"
                },
                "table_schema_rename": {
                    "type": "string"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TableConfig"
                    }
                }
            }
        },
        "models.DatabaseConnectionConfig": {
            "type": "object",
            "required": [
                "database_type",
                "host",
                "password",
                "port",
                "user"
            ],
            "properties": {
                "database_type": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "models.DestTaskConfig": {
            "type": "object",
            "required": [
                "connection_config",
                "task_name"
            ],
            "properties": {
                "connection_config": {
                    "$ref": "#/definitions/models.DatabaseConnectionConfig"
                },
                "database_type": {
                    "type": "string"
                },
                "mysql_dest_task_config": {
                    "$ref": "#/definitions/models.MysqlDestTaskConfig"
                },
                "node_id": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
...
        "models.GtidModeValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is a string version of any error that may have occured",
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "models.MysqlDestTaskConfig": {
            "type": "object",
            "properties": {
                "dependency_history_size": {
                    "type": "integer"
                },
                "parallel_workers": {
                    "type": "integer"
                },
                "use_my_sql_dependency": {
                    "type": "boolean"
                }
            }
        },
        "models.MysqlSrcTaskConfig": {
            "type": "object",
            "properties": {
                "auto_gtid": {
                    "type": "boolean"
                },
                "binlog_relay": {
                    "type": "boolean"
                },
                "expand_syntax_support": {
                    "type": "boolean"
                },
                "gtid": {
                    "type": "string"
                },
                "wait_on_job": {
                    "type": "string"
                }
            }
        },
        "models.MysqlTaskValidationReport": {
            "type": "object",
            "properties": {
                "binlog_validation": {
                    "$ref": "#/definitions/models.BinlogValidation"
                },
                "connection_validation": {
                    "$ref": "#/definitions/models.ConnectionValidation"
                },
                "gtid_mode_validation": {
                    "$ref": "#/definitions/models.GtidModeValidation"
                },
                "privileges_validation": {
                    "$ref": "#/definitions/models.PrivilegesValidation"
                },
                "server_id_validation": {
                    "$ref": "#/definitions/models.ServerIDValidation"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
...
        "models.OracleSrcTaskConfig": {
            "type": "object",
            "properties": {
                "scn": {
                    "type": "integer"
                }
            }
        },
...
        "models.PrivilegesValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is a string version of any error that may have occured",
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "models.ServerIDValidation": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is a string version of any error that may have occured",
                    "type": "string"
                },
                "validated": {
                    "type": "boolean"
                }
            }
        },
...
        "models.SrcTaskConfig": {
            "type": "object",
            "required": [
                "connection_config",
                "task_name"
            ],
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "connection_config": {
                    "$ref": "#/definitions/models.DatabaseConnectionConfig"
                },
                "drop_table_if_exists": {
                    "type": "boolean"
                },
                "group_max_size": {
                    "type": "integer"
                },
                "group_timeout": {
                    "type": "integer"
                },
                "mysql_src_task_config": {
                    "$ref": "#/definitions/models.MysqlSrcTaskConfig"
                },
                "node_id": {
                    "type": "string"
                },
                "oracle_src_task_config": {
                    "$ref": "#/definitions/models.OracleSrcTaskConfig"
                },
                "repl_chan_buffer_size": {
                    "type": "integer"
                },
                "replicate_do_db": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DataSourceConfig"
                    }
                },
                "replicate_ignore_db": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DataSourceConfig"
                    }
                },
                "skip_create_db_table": {
                    "type": "boolean"
                },
                "sql_filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
        "models.TableConfig": {
            "type": "object",
            "properties": {
                "column_map_from": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "table_name": {
                    "type": "string"
                },
                "table_regex": {
                    "type": "string"
                },
                "table_rename": {
                    "type": "string"
                },
                "where": {
                    "type": "string"
                }
            }
        },
...
        "models.ValidateJobReqV2": {
            "type": "object",
            "required": [
                "dest_task",
                "job_id",
                "src_task"
            ],
            "properties": {
                "dest_task": {
                    "$ref": "#/definitions/models.DestTaskConfig"
                },
                "is_password_encrypted": {
                    "type": "boolean"
                },
                "job_id": {
                    "type": "string"
                },
                "src_task": {
                    "$ref": "#/definitions/models.SrcTaskConfig"
                }
            }
        },
        "models.ValidateJobRespV2": {
            "type": "object",
            "properties": {
                "driver_config_validated": {
                    "description": "DriverConfigValidated indicates whether the agent validated the driver",
                    "type": "boolean"
                },
                "job_validation_error": {
                    "type": "string"
                },
                "job_validation_warning": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "mysql_task_validation_report": {
                    "description": "config",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MysqlTaskValidationReport"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}